% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-soiltexture2-classes.R
\name{createTextureSystem}
\alias{createTextureSystem}
\title{Creates a textureSystem-class object (the full definition of texture plot and its optionally classification).}
\usage{
createTextureSystem(ternarySystem, particleSizeBoundaries, ...)
}
\arguments{
\item{ternarySystem}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{particleSizeBoundaries}{See \code{\link[soiltexture2]{textureSystem-class}}.}

\item{\dots}{Additional parameters passed to 
\code{\link[ternaryplot]{ternaryCheck}}}
}
\value{
A \code{\link[soiltexture2]{textureSystem-class}}.
}
\description{
Creates a \code{\link[soiltexture2]{textureSystem-class}} 
 object (the full definition of texture plot and its 
 optionally classification).
}
\examples{

library( "soiltexture2" ) 

s <- createTernarySystem( 
    "ternaryVariables" = createTernaryVariables( 
        "blrNames"  = c( "SAND", "CLAY", "SILT" ), 
        "blrLabels" = c(
            "Sand 0.05 - 2 mm [\%]", 
            "Clay 0 - 0.002 mm [\%]", 
            "Silt 0.002 - 0.05 mm [\%]" 
        ) ),  
    "ternaryGeometry" = createTernaryGeometry( ##
        "fracSum"   = 100, 
        "tlrAngles" = c( 60, 60, 60 ), # top, left, right angles
        "blrClock"  = c( TRUE, TRUE, TRUE )
    ),  
    "main" = "My texture plot" ) 

psb <- c( 0, "CLAY" = 2, "SILT" = 50, "SAND" = 2000 ) 

st <- createTextureSystem(
    "ternarySystem"          = s, 
    "particleSizeBoundaries" = psb )       

psb

#   ternaryPlot( st )
}
\seealso{
\code{\link[soiltexture2]{textureSystem-class}}.
}

