% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa04-textureSystems.R
\name{getAllTextureSystems}
\alias{getAllTextureSystems}
\title{INTERNAL: Fetch all the pre-defined texture classification systems}
\usage{
getAllTextureSystems()
}
\value{
A \code{\link[base]{list}} 
 \code{\link[soiltexture2]{textureSystem-class}} objects.
}
\description{
INTERNAL: Fetch all the pre-defined texture classification 
 systems, as a list of 
 \code{\link[soiltexture2]{textureSystem-class}}es.
}
\details{
The function is designed for internal use, in combination 
 with the argument \code{terSysEnvList} of the functions 
 \code{link[soiltexture2]{stPar}} or 
 \code{link[soiltexture2]{getStPar}}, as a flexible 
 mechanism to manage the list of 
 \code{\link[soiltexture2]{textureSystem-class}} objects 
 used by the package.
}
\keyword{internal}

