% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa01-soiltexture2-options.R
\name{getStPar}
\alias{getStPar}
\title{Get a single default parameters for the package.}
\usage{
getStPar(par)
}
\arguments{
\item{par}{See the \code{par} argument in \code{\link{stPar}}. Notice that if
more than one parameter name is provided, only the first one will be
returned.}
}
\value{
Return the value of the parameter \code{par}, without the list
 container of \code{\link{stPar}}.
}
\description{
Get a single default parameters for the package. Wrapper around
 \code{\link{stPar}}.
}

