% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa03-soiltexture2-classes-utility.R
\name{print.textureSystem}
\alias{print.textureSystem}
\title{Print the content of a textureSystem object in a human readable format.}
\usage{
\method{print}{textureSystem}(x, prefix = "", collapse = "; ", ...)
}
\arguments{
\item{x}{A \code{\link[soiltexture2]{textureSystem-class}} object, 
as created with \code{\link[soiltexture2]{createTextureSystem}}.}

\item{prefix}{Single character string. Prefix used before the different 
items in \code{x} (intended for internal use, for example 
\code{prefix = "$ternarySystem"}).}

\item{collapse}{Single character string. Passed to 
\code{\link{paste}( ..., collapse )} when displaying the 
items' values.}

\item{\dots}{Additional parameters passed to specific methods (not 
used).}
}
\description{
Print the content of a 
 \code{\link[soiltexture2]{textureSystem-class}} object 
 (S3-class) in a human readable format.
}

