% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa00-soiltexture2-package.R
\docType{package}
\name{soiltexture2-package}
\alias{soiltexture2-package}
\title{Utility functions for soil texture ternary plots and soil texture classification.}
\description{
Utility functions for soil texture ternary plots and soil texture classification.
}
\details{
Set of R functions designed to produce soil texture
 triangles (also called soil texture plots, soil
 texture diagrams or soil texture ternary plots),
 classify and transform soil textures data. These
 functions allows to plot any soil texture triangle
 (classification) into different triangle geometries
 (isosceles, right-angled triangles, etc.). This set
 of function is expected to be useful to people using
 soil textures data from different soil texture
 classification or different particle size systems.
 Many (> 15) texture triangles from all around the
 world will be predefined in the package. 'soiltexture2'
 is a complete redesign and re-documentation of the
 soiltexture-package, from which is keeps the core features.
}
\examples{
 # Examples coming later

}
\author{
Julien Moeys \email{Julien.Moeys@slu.se} 

Maintainer: Julien Moeys \email{Julien.Moeys@slu.se}
}
\keyword{package}

