% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa01-soiltexture2-options.R
\docType{data}
\name{.stParList}
\alias{.stParList}
\title{Environment containng a backup copy of all soiltexture2 package-options}
\format{An environment, where each item corresponds to an 
 argument in \code{\link[soiltexture2]{stPar}}.}
\usage{
.stParList
}
\description{
Environment containing a backup copy of all soiltexture2 
 package-options
}
\keyword{internal}

