% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-soiltexture2-classes.R
\name{textureSystem-class}
\alias{textureSystem-class}
\title{Class to hold the definition of a textureSystem (the full definition of texture plot and optionally classification).}
\description{
Class (\code{S3}) to hold the definition of a textureSystem 
 (the full definition of texture plot and optionally 
 classification). A \code{textureSystem-class} object is 
 a tagged \code{\link[base]{list}} of several 
 objects/items (see below). A 'tagged' list means a list of 
 the form \code{list( "tag1" = value1, "tag2" = value2, 
 ... )}.
}
\details{
The \code{S3} \code{\link[base]{class}} system is an 
 \emph{in}formal and lightweight class system.
}
\section{List objects/items}{


 A \code{textureSystem-class} object has all the features 
 (items) of a A \code{\link[ternaryplot]{ternarySystem-class}} 
 object, plus the following item(s): 

\itemize{
 \item \bold{\code{particleSizeBoundaries}}: A tagged vector 
   of numeric or integer values giving the boundaries of 
   the particle size classes. The vector must be ordered 
   and increasing. The length of the vector must be 4, 
   with the 1st value giving the (absolute) lowest value 
   of the particle size, and the 3 remaining values the 
   upper boundary of each texture particle size class.
   The tags of the 1st item can be any value (even 
   \code{""}) and the 3 remaining tags \bold{must} 
   correspond to the variable-name of the 3 texture 
   particle size classes (as can be obtained with 
   \code{blrNames(s)}, where \code{s} is the 
   \code{textureSystem-class} object, but the order of does 
   not has to be the same of course). Finally, the 
   \bold{unit} should (preferably) be in micrometers. 
   
   A valid example could be \code{c(0,"CLAY"=2,"SILT"=50,"SAND"=2000)}.
}
}
\seealso{
A \code{textureSystem-class} object is compatible with 
 all the functions available for 
 \code{\link[ternaryplot]{ternarySystem-class}} (see 'See 
 also' in that help page). In addition to that, the following 
 functions can be relevant:

\itemize{
 \item TO BE COMPLETED.
}
}

